Fix bug with watch list ui state and clean up files
