+++ ios/Cut/Account/SessionManager.swift
