+++ ios/Cut/UI/Color.swift
@@ -1,0 +1,14 @@
+//
+//  Color.swift
+//  Cut
+//
+//  Created by Kyle Satti on 2/22/24.
+//
+
+import SwiftUI
+
+extension Color {
+    static var sub: Color {
+        return Color(white: 0, opacity: 0.2)
+    }
+}
