+++ ios/Cut/UI/ImageStack.swift
@@ -1,0 +1,44 @@
+//
+//  ImageStack.swift
+//  Cut
+//
+//  Created by Kyle Satti on 2/22/24.
+//
+
+import SwiftUI
+
+struct ImageStack: View {
+    let urls: [String]
+    private let spacing: CGFloat = 4
+
+    var body: some View {
+        HStack(spacing: -spacing) {
+            ForEach(0..<urls.count, id: \.self) { i in
+                if i < urls.count - 1 {
+                    Circle()
+                        .foregroundColor(.red)
+                        .mask(MoonMask(amount: spacing).fill(style: FillStyle(eoFill: true)))
+                } else {
+                    Circle()
+                        .foregroundColor(.red)
+                }
+            }
+        }
+        .frame(maxHeight: spacing * 4)
+    }
+}
+
+struct MoonMask: Shape {
+    let amount: CGFloat
+
+    func path(in rect: CGRect) -> Path {
+        var path = Rectangle().path(in: rect)
+        path.addPath(Circle().path(in: rect)
+            .offsetBy(dx: amount * 2.4, dy: 0))
+        return path
+    }
+}
+
+#Preview {
+    return ImageStack(urls: ["test", "rest"])
+}
